requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aiohttp-swagger==1.0.15
aiojobs==0.3.0
aiopg==1.3.1
ansible==2.9.27
beautifulsoup4==4.10.0
netifaces==0.11.0
nltk==3.6.3
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.25